corrected remarks and add class card and configcard
